♻️(summary) reorganize API code
I totally reorganized the API code, to gain in clarity.
