+++ src/helm/env.d/staging/values.meet.yaml.gotmpl
@@ -280,7 +280,7 @@ celery:
   command:
     - "celery"
     - "-A"
-    - "summary.celery_worker"
+    - "summary.core.celery_worker"
     - "worker"
     - "--pool=solo"
     - "--loglevel=info"
