+++ src/summary/summary/api/__init__.py
@@ -1,0 +1,1 @@
+"""Summary Api package."""
