+++ src/summary/summary/api/route/__init__.py
@@ -1,0 +1,1 @@
+"""Summary Api Route package."""
