+++ src/summary/summary/core/__init__.py
@@ -1,0 +1,1 @@
+"""Summary Core package."""
