+++ src/summary/summary/main.py
@@ -1,46 +1,16 @@
-"""Application endpoint."""
+"""Application."""
 
-from celery.result import AsyncResult
-from fastapi import Depends, FastAPI
-from pydantic import BaseModel
+from fastapi import FastAPI
 
-from .celery_worker import process_audio_transcribe_summarize
-from .security import verify_token
+from summary.api import health
+from summary.api.main import api_router
+from summary.core.config import get_settings
 
-app = FastAPI()
+settings = get_settings()
 
+app = FastAPI(
+    title=settings.app_name,
+)
 
-@app.get("/__heartbeat__")
-async def heartbeat():
-    """Health check endpoint for monitoring."""
-    return {"status": "ok"}
-
-
-@app.get("/__lbheartbeat__")
-async def lbheartbeat():
-    """Health check endpoint for load balancer."""
-    return {"status": "ok"}
-
-
-class NotificationRequest(BaseModel):
-    """Notification data."""
-
-    filename: str
-    email: str
-    sub: str
-
-
-@app.post("/push")
-async def notify(request: NotificationRequest, token: str = Depends(verify_token)):
-    """Push a notification."""
-    task = process_audio_transcribe_summarize.delay(
-        request.filename, request.email, request.sub
-    )
-    return {"task_id": task.id, "message": "Notification sent"}
-
-
-@app.get("/status/{task_id}")
-async def get_status(task_id: str, token: str = Depends(verify_token)):
-    """Check task status by ID."""
-    task = AsyncResult(task_id)
-    return {"task_id": task_id, "status": task.status}
+app.include_router(api_router, prefix=settings.app_api_v1_str)
+app.include_router(health.router)
